Refactored event_finder_rhodesmill and tests/event_finder/ files to move changes from tests to main files. Modified tests for test_event_finder.py and test_azimuth_filtering.py
