+++ satellite_determination/azimuth_filter/azimuth_filtering.py
@@ -1,14 +1,15 @@
-from datetime import datetime, timedelta
+from datetime import timedelta
 from typing import List
 
 import pytz
-from skyfield.api import wgs84
+from skyfield.toposlib import wgs84
+
+from satellite_determination.azimuth_filter.overhead_window_from_azimuth import OverheadWindowFromAzimuth
 from satellite_determination.custom_dataclasses.overhead_window import OverheadWindow
 from satellite_determination.custom_dataclasses.reservation import Reservation
-from satellite_determination.custom_dataclasses.time_window import TimeWindow
-from tests.validator.test_overhead_window_from_azimuth import OverheadWindowFromAzimuth
 
-class TestAzimuthFiltering:
+
+class AzimuthFilter:
 
     def __init__(self, overhead_windows: List[OverheadWindow], reservation: Reservation):
         self._overhead_windows = overhead_windows
@@ -18,9 +19,12 @@ class TestAzimuthFiltering:
         azimuth_filtered_overhead_windows = []
         coordinates = wgs84.latlon(self._reservation.facility.point_coordinates.latitude,
                                    self._reservation.facility.point_coordinates.longitude)
-        time_delta = timedelta(minutes=1) #timedelta to check new azimuth
+        time_delta = timedelta(seconds=1) #timedelta to check new azimuth, if we move to more granular seconds takes FOREVER to run but get more windows
         for window in self._overhead_windows:
-            difference = window.satellite - coordinates  # get vector sat relative to telescope location
+            satellite_azimuth_values = []
+            satellite_time_values = []
+            difference = window.satellite - coordinates
+            #difference = window.satellite.to_rhodesmill() - coordinates  # get vector sat relative to telescope location
             while window.overhead_time.begin.astimezone(pytz.utc) < window.overhead_time.end.astimezone(pytz.utc):  # timestamp for events needs to be changed since when the sat enters cone is not necessarily event timestamp
                 topocentric = difference.at(window.overhead_time.begin)
                 alt, az, distance = topocentric.altaz()
