+++ satellite_determination/custom_dataclasses/frequency_range.py
@@ -1,5 +1,7 @@
 from dataclasses import dataclass
 from enum import Enum
+from typing import List
+import csv
 
 
 class FrequencyRangeJsonKey(Enum):
@@ -9,11 +11,10 @@ class FrequencyRangeJsonKey(Enum):
 
 @dataclass
 class FrequencyRange:
-    high_in_megahertz: float
-    low_in_megahertz: float
+    frequencies: List[tuple]
 
     @classmethod
-    def from_json(cls, info: dict) -> 'FrequencyRange':
+    def from_csv(cls, info: dict) -> 'FrequencyList':
         return cls(
             high_in_megahertz=info[FrequencyRangeJsonKey.high_in_megahertz.name],
             low_in_megahertz=info[FrequencyRangeJsonKey.low_in_megahertz.name]
