+++ satellite_determination/event_finder/event_finder_rhodesmill/support/overhead_window_from_events.py
