+++ satellite_determination/event_finder/validator.py
