+++ satellite_determination/window_finder.py
@@ -6,7 +6,7 @@ from satellite_determination.custom_dataclasses.overhead_window import OverheadW
 from satellite_determination.custom_dataclasses.reservation import Reservation
 from satellite_determination.custom_dataclasses.satellite.satellite import Satellite
 from satellite_determination.custom_dataclasses.time_window import TimeWindow
-from satellite_determination.validator.validator import Validator
+from satellite_determination.event_finder.validator import Validator
 
 
 @dataclass
