+++ tests/event_finder/satellite_reference_file
