--- tests/validator/test_validator.py
@@ -1,76 +1,0 @@
-import pytz
-from skyfield.api import load, wgs84
-import datetime
-import os
-from pathlib import Path
-from datetime import datetime, timedelta
-import filecmp
-from typing import List
-
-from satellite_determination.custom_dataclasses.coordinates import Coordinates
-from satellite_determination.custom_dataclasses.facility import Facility
-from satellite_determination.custom_dataclasses.reservation import Reservation
-from satellite_determination.retrievers.satellite_retriever.skyfield_satellite_retriever import SkyfieldSatelliteList
-from satellite_determination.custom_dataclasses.time_window import TimeWindow
-from satellite_determination.utilities import convert_dt_to_utc
-from tests.utilities import get_script_directory
-from satellite_determination.validator.validator_skyfield.support.overhead_window_from_events import \
-    EventTypesRhodesmill, EventRhodesmill, OverheadWindowFromEvents
-from tests.validator.test_azimuth_filtering import TestAzimuthFiltering
-
-
-class ValidatorRhodesMill:
-
-    def __init__(self, list_of_satellites: SkyfieldSatelliteList, reservation: Reservation):
-        self._list_of_satellites = list_of_satellites
-        self._reservation = reservation
-
-    def get_overhead_windows(self):
-        ts = load.timescale()
-        overhead_windows = []
-        t0 = ts.from_datetime(convert_dt_to_utc(self._reservation.time.begin))
-        t1 = ts.from_datetime(convert_dt_to_utc(self._reservation.time.end))
-        coordinates = wgs84.latlon(self._reservation.facility.point_coordinates.latitude, self._reservation.facility.point_coordinates.longitude)
-        for sat in self._list_of_satellites.satellites:
-            t, events = sat.find_events(coordinates, t0, t1, altitude_degrees=self._reservation.facility.angle_of_visibility_cone)
-            if events.size == 0:
-                continue
-            else:
-                rhodesmill_event_list = []
-                for ti, event in zip(t, events):
-                    if event == 0:
-                        translated_event = EventRhodesmill(event_type=EventTypesRhodesmill.ENTERS, satellite=sat, timestamp=ti)
-                    elif event == 1:
-                        translated_event = EventRhodesmill(event_type=EventTypesRhodesmill.CULMINATES, satellite=sat, timestamp=ti)
-                    elif event == 2:
-                        translated_event = EventRhodesmill(event_type=EventTypesRhodesmill.EXITS, satellite=sat, timestamp=ti)
-                    rhodesmill_event_list.append(translated_event)
-                sat_windows = OverheadWindowFromEvents(events=rhodesmill_event_list, reservation=self._reservation).get()
-                for window in sat_windows:
-                    overhead_windows.append(window)
-        azimuth_filtered_windows = TestAzimuthFiltering(overhead_windows=overhead_windows, reservation=self._reservation).filter_azimuth()
-        return azimuth_filtered_windows
-        #return overhead_windows
-
-
-class TestWindowListFinder:
-
-    def test_get_window_list(self):
-        tle_file = Path(get_script_directory(__file__), 'TLEdata', 'test.txt')
-        list_of_satellites = SkyfieldSatelliteList.load_tle(str(tle_file))
-        print(len(list_of_satellites.satellites))
-        for sat in list_of_satellites.satellites:
-            print(sat)
-        reservation = Reservation(facility=Facility(angle_of_visibility_cone=0, point_coordinates=Coordinates(latitude=40.8178049, longitude=-121.4695413),name='name', azimuth=30),
-                                  time=TimeWindow(begin=datetime(year=2023, month=2, day=22, hour=1, tzinfo=pytz.utc), end=datetime(year=2023, month=2, day=22, hour=2, tzinfo=pytz.utc)))
-        overhead_windows = ValidatorRhodesMill(list_of_satellites=list_of_satellites, reservation=reservation).get_overhead_windows()
-        print(len(overhead_windows))
-        #with open ("satellite_overhead_test", "w") as outfile:
-            #outfile.writelines(str(overhead_windows))
-            #outfile.close()
-
-        #assert filecmp.cmp('./tests/validator/overhead_window_reference.txt', 'satellite_overhead_test') == 1
-        #os.remove("satellite_overhead_test")
-
-blah = TestWindowListFinder()
-blah.test_get_window_list()
